+++ lib/docker_swarm/models/service.rb
@@ -8,5 +8,14 @@ module DockerSwarm
     include Concerns::Updatable
     include Concerns::Deletable
     include Concerns::Loggable
+
+    # Restarts the service by incrementing ForceUpdate, which causes
+    # Docker to recreate all tasks.
+    #
+    # @return [Boolean] true if the restart was triggered successfully
+    def restart
+      current = self.Spec&.dig("TaskTemplate", "ForceUpdate").to_i
+      update(Spec: { TaskTemplate: { ForceUpdate: current + 1 } })
+    end
   end
 end